requirements.txt
strands-agents
strands-agents-tools
bedrock-agentcore-starter-toolkit
boto3